For 4596: collection interactor rename + kdoc
