+++ app/src/main/java/org/mozilla/fenix/collections/SaveCollectionListAdapter.kt
@@ -9,7 +9,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import io.reactivex.Observer
 import kotlinx.android.synthetic.main.collections_list_item.view.*
 import org.mozilla.fenix.R
 import org.mozilla.fenix.components.description
@@ -18,7 +17,7 @@ import org.mozilla.fenix.home.sessioncontrol.Tab
 import org.mozilla.fenix.home.sessioncontrol.TabCollection
 
 class SaveCollectionListAdapter(
-    private val interactor: CollectionViewInteractor
+    private val interactor: CollectionCreationInteractor
 ) : RecyclerView.Adapter<CollectionViewHolder>() {
 
     private var tabCollections = listOf<TabCollection>()
